Split settings between test and live
In order to be able to easily separate the settings
between the test and live environment, I'm creating
a settings module that automatically loads the
settings file for the correct environment.

Note that the settings now need a local configuration
file. This is not included in this commit, and
must be added manually by the developer.